TriggerDataDecoder and MRDTracks minor bug fix (#292)
* Update FindMrdTracks.cpp

Don't hault the toolchain if there are no MRD tracks

* Update TriggerDataDecoder.cpp

Triggers are stored to the buffer - clear it to speed up the tool. Also adding verbosity conditions on std out

* Update LoadRawDataConfig

Wrong config variables - now it should work out of the box as intended

* Update CreateMyList.sh

Rather than passing the path of the RAWData file to the bash script, add it hard coded. Not great practice but this script is used for a very specific purpose to generate a list file containing part files from the RAWData, and the data location has not changed in years (and likely won't change for now).

* Update FindMrdTracks.cpp

revert changes